+++ taxsystem/api/taxsystem/corporation.py
@@ -28,12 +28,12 @@ class CorporationApiEndpoints:
             tags=self.tags,
         )
         def get_payments(request, corporation_id: int):
-            corp, perms = get_manage_corporation(request, corporation_id)
+            owner, perms = get_manage_corporation(request, corporation_id)
 
-            if corp is None:
+            if owner is None:
                 return 404, "Corporation Not Found"
 
-            payments = Payments.objects.filter(account__corporation=corp)
+            payments = Payments.objects.filter(account__owner=owner)
 
             payments_dict = {}
 
@@ -74,11 +74,9 @@ class CorporationApiEndpoints:
             if corp is None:
                 return 404, "Corporation Not Found"
 
-            account = PaymentSystem.objects.get(corporation=corp, user=request.user)
+            account = PaymentSystem.objects.get(owner=corp, user=request.user)
 
-            payments = Payments.objects.filter(
-                account__corporation=corp, account=account
-            )
+            payments = Payments.objects.filter(account__owner=corp, account=account)
 
             own_payments_dict = {}